requirements.txt
streamlit
torch
torchvision
opencv-python
matplotlib
pillow
